[feature/enhance-highways-reststop] 경로 내 휴게소 로직 수정
#27

+++ core/src/main/kotlin/com/eightpotatoes/nexters/core/model/Location.kt
@@ -1,3 +1,10 @@
 package com.eightpotatoes.nexters.core.model
 
-data class Location(val latitude: Float, val longitude: Float)
+/**
+ * Location 위도 경도
+ *
+ * @property latitude 위도 (37.123456)
+ * @property longitude 경도 (127.123456)
+ * @constructor Create Location
+ */
+data class Location(val latitude: Double, val longitude: Double)
